divide clipboard commands from element_commands
